Student Interface Update -0.3
I have progressed the database interface on a single view:
/students/listclusters to the point the controller is capable of listing
all the correct information, and the view is able to process them.
Although at this time with only one view finished there is no
functionality. However this will be rapidly developed with a framework I
can now develop from.

PLEASE NOTE:
Resolved issues have been raised in the GitHub Issue Log in relation to
SQL Database creation. These will need to be read in the event any
interaction with the database is required.